Move util function to util folder
